requirements.txt
annotated-types==0.5.0
chromadb==0.4.4
httptools==0.6.0
openai==0.27.8
pip-chill==1.0.3
pydantic-core==2.4.0
pypdf==3.14.0
python-dotenv==1.0.0
pyyaml==6.0.1
replicate==0.10.0
termcolor==2.3.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
